tests: Use Ginkgo labels for consistency with other projects
[#181108660](https://www.pivotaltracker.com/story/show/181108660)
