+++ acceptance-tests/README.md
@@ -27,5 +27,5 @@ The main pattern is:
 
 ### Run tests
 
-- Move in the test folder e.g `acceptance-tests/redis`
+- Move in the test folder e.g `acceptance-tests`
 - Run tests from IDE or run command `go run github.com/onsi/ginkgo/v2/ginkgo -v`
