+++ acceptance-tests/acceptance_tests_suite_test.go
@@ -1,0 +1,20 @@
+package acceptance_test
+
+import (
+	"testing"
+	"csbbrokerpakgcp/acceptance-tests/helpers/environment"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestAcceptanceTests(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Acceptance Tests Suite")
+}
+
+var GCPMetadata environment.GCPMetadata
+
+var _ = BeforeSuite(func() {
+	GCPMetadata = environment.ReadGCPMetadata()
+})
