+++ acceptance-tests/errormessages_test.go
@@ -1,4 +1,4 @@
-package errormessages_test
+package acceptance_test
 
 import (
 	"csbbrokerpakgcp/acceptance-tests/helpers/brokers"
@@ -12,7 +12,7 @@ import (
 	. "github.com/onsi/gomega/gexec"
 )
 
-var _ = Describe("Error Messages", func() {
+var _ = Describe("Error Messages", Label("errormessages"), func() {
 	When("the create-service command fails immediately", func() {
 		It("prints a useful error message", func() {
 			name := random.Name(random.WithPrefix("error"))
