+++ acceptance-tests/helpers/apps/testapps.go
@@ -1,6 +1,9 @@
 package apps
 
-import "fmt"
+import (
+	"fmt"
+	"os"
+)
 
 type AppCode string
 
@@ -15,7 +18,15 @@ const (
 )
 
 func (a AppCode) Dir() string {
-	return fmt.Sprintf("../apps/%s", string(a))
+	for _, d := range []string{"apps", "../apps"} {
+		p := fmt.Sprintf("%s/%s", d, string(a))
+		_, err := os.Stat(p)
+		if err == nil {
+			return p
+		}
+	}
+
+	panic(fmt.Sprintf("could not find source for app: %s", a))
 }
 
 func WithApp(app AppCode) Option {
