--- acceptance-tests/mysql/mysql_suite_test.go
@@ -1,13 +1,0 @@
-package mysql_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestMysql(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Mysql Suite")
-}
