--- acceptance-tests/postgresql/postgresql_suite_test.go
@@ -1,13 +1,0 @@
-package postgresql_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestPostgresql(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "PostgreSQL Suite")
-}
