--- acceptance-tests/redis/redis_suite_test.go
@@ -1,13 +1,0 @@
-package redis_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestRedis(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Redis Suite")
-}
