--- acceptance-tests/spanner/spanner_suite_test.go
@@ -1,13 +1,0 @@
-package spanner_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestSpanner(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Spanner Suite")
-}
