+++ acceptance-tests/spanner_test.go
@@ -1,4 +1,4 @@
-package dataproc_test
+package acceptance_test
 
 import (
 	"csbbrokerpakgcp/acceptance-tests/helpers/apps"
@@ -10,14 +10,14 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Dataproc", func() {
+var _ = Describe("Spanner", Label("spanner"), func() {
 	It("can be accessed by an app", func() {
 		By("creating a service instance")
-		serviceInstance := services.CreateInstance("csb-google-dataproc", "standard")
+		serviceInstance := services.CreateInstance("csb-google-spanner", "small")
 		defer serviceInstance.Delete()
 
 		By("pushing the unstarted app")
-		appOne := apps.Push(apps.WithApp(apps.Dataproc))
+		appOne := apps.Push(apps.WithApp(apps.Spanner))
 		defer apps.Delete(appOne)
 
 		By("binding the app to the service instance")
@@ -29,15 +29,13 @@ var _ = Describe("Dataproc", func() {
 		By("checking that the app environment has a credhub reference for credentials")
 		Expect(binding.Credential()).To(matchers.HaveCredHubRef)
 
-		By("running a job")
-		jobName := random.Hexadecimal()
-		appOne.PUT("", jobName)
+		By("setting a key-value using the app")
+		key := random.Hexadecimal()
+		value := random.Hexadecimal()
+		appOne.PUT(value, key)
 
-		By("getting the job status")
-		status := appOne.GET(jobName)
-		Expect(status).To(Equal("DONE"))
-
-		By("deleting the job")
-		appOne.DELETE(jobName)
+		By("getting the value using the same app")
+		got := appOne.GET(key)
+		Expect(got).To(Equal(value))
 	})
 })
