+++ acceptance-tests/storage_test.go
@@ -1,4 +1,4 @@
-package storage_test
+package acceptance_test
 
 import (
 	"csbbrokerpakgcp/acceptance-tests/helpers/apps"
@@ -10,7 +10,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Storage", func() {
+var _ = Describe("Storage", Label("storage"), func() {
 	It("can be accessed by an app", func() {
 		By("creating a service instance")
 		serviceInstance := services.CreateInstance("csb-google-storage-bucket", "private")
