--- acceptance-tests/withoutcredhub/withoutcredhub_suite_test.go
@@ -1,13 +1,0 @@
-package withoutcredhub_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestWithoutCredHub(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Without CredHub")
-}
